refactor: 프로젝트 파일 구조 정리 및 이름 변경
