+++ src/model/StringCalculator.js
